@@ -17,7 +17,7 @@ class MyStreamEventProducer : () -> Flux<Message<MyEventPayload>>, MyEventProduc
   override fun produce(event: MyEvent) {
     val message = MessageBuilder
       .withPayload(toPayload(event))
-      .setHeader(KafkaHeaders.MESSAGE_KEY, toKey(event))
+      .setHeader(KafkaHeaders.KEY, toKey(event))
       .build()
     sink.emitNext(message, FAIL_FAST)
   }
